deps: update discord.js
