+++ package.json
@@ -28,11 +28,11 @@
     "url": "https://github.com/DTrombett/ms-bot/issues"
   },
   "dependencies": {
-    "@discordjs/builders": "^0.13.0-dev.1646913825.dcd4797",
+    "@discordjs/builders": "^0.13.0-dev.1647259738.2297c2b",
     "@discordjs/rest": "^0.4.0-dev.1647259762.2297c2b",
     "bufferutil": "^4.0.6",
     "discord-api-types": "*",
-    "discord.js": "^14.0.0-dev.1646914074.dcd4797",
+    "discord.js": "^14.0.0-dev.1647259751.2297c2b",
     "erlpack": "github:discord/erlpack",
     "prettier": "^2.6.0",
     "typescript": "^4.6.2",
